pyproject.toml
[tool.poetry]
name = "SHIFT"
version = "0.1.0"
description = ""
authors = ["User <sharapov-kirill@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.30"
pydantic-settings = "^2.2.1"
pydantic = {extras = ["email"], version = "^2.7.1"}
python-dotenv = "^1.0.1"
bcrypt = "^4.1.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
alembic = "^1.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
